feat(groups): revamp todo-comments colors
added colors for the missing todos (PERF, WARN, NOTE, and TEST), and
changed some of the colors around. most notably, changed the TODO color
from green to grey (the same color of a regular comment). since this is
the most common todo-comment people are likely to make, it's better for
it to be unobstrusive